Add DEBUG and Optimal Strategy mode
Adds a debug mode where upcoming cards and dealers hidden card
are shown.

Also adds a optimal strategy mode, which runs x hands locally
and calculates the optimal action (HIT or STAND) for the player.

DEBUG flag will be added to settings menu when that exists in the
future, now it's just a local bool.

BlackJackScreen needs some refactor, it's a bit messy.
Also some weird recompositions happening when simulation runs,
needs to be fixed.